refactor: rearrange structure of crypto backends
