+++ didcomm_messaging/crypto/backend/basic.py
@@ -1,7 +1,7 @@
 """Basic Crypto Implementations."""
 
 from typing import Optional
-from . import S, SecretsManager
+from .base import S, SecretsManager
 
 
 class InMemorySecretsManager(SecretsManager[S]):
